Move attachpoint checking out of semantic_analyser
Move into a separate pass in attachpoint_parser,
which gets renamed to 'attachpoint_passes'.
Created two separate passes because there are a lot
of tests that have bad attachpoints (e.g. `k:asdf { 1 }`)
so we just don't add the attachpoint checking
pass for these tests.

This is part of the larger effor to slim down
semantic_analyser.

Also added a few more unit tests for this attachpoint
checking functionality.


stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4650, branch: user/jordalgo/semantic_analyser_refactor/1
