+++ src/ast/passes/attachpoint_passes.cpp
@@ -6,10 +6,12 @@
 #include <string>
 #include <vector>
 
+#include "arch/arch.h"
 #include "ast/ast.h"
-#include "ast/attachpoint_parser.h"
 #include "ast/context.h"
 #include "ast/helpers.h"
+#include "ast/passes/attachpoint_passes.h"
+#include "ast/visitor.h"
 #include "types.h"
 #include "util/int_parser.h"
 #include "util/paths.h"
@@ -18,6 +20,309 @@
 #include "util/wildcard.h"
 
 namespace bpftrace::ast {
+
+class AttachPointChecker : public Visitor<AttachPointChecker> {
+public:
+  explicit AttachPointChecker(ASTContext &ast,
+                              BPFtrace &bpftrace,
+                              bool listing,
+                              bool has_child = true)
+      : ast_(ast),
+        bpftrace_(bpftrace),
+        listing_(listing),
+        has_child_(has_child) {};
+
+  using Visitor<AttachPointChecker>::visit;
+  void visit(AttachPoint &ap);
+
+private:
+  ASTContext &ast_;
+  BPFtrace &bpftrace_;
+  bool listing_;
+  bool has_begin_probe_ = false;
+  bool has_end_probe_ = false;
+  bool has_child_ = false;
+  std::unordered_map<std::string, Location> benchmark_locs_;
+};
+
+void AttachPointChecker::visit(AttachPoint &ap)
+{
+  if (ap.provider == "kprobe" || ap.provider == "kretprobe") {
+    if (ap.func.empty())
+      ap.addError() << "kprobes should be attached to a function";
+    // Warn if user tries to attach to a non-traceable function
+    if (bpftrace_.config_->missing_probes != ConfigMissingProbes::ignore &&
+        !util::has_wildcard(ap.func) && !bpftrace_.is_traceable_func(ap.func)) {
+      ap.addWarning() << ap.func
+                      << " is not traceable (either non-existing, inlined, "
+                         "or marked as "
+                         "\"notrace\"); attaching to it will likely fail";
+    }
+  } else if (ap.provider == "uprobe" || ap.provider == "uretprobe") {
+    if (ap.target.empty())
+      ap.addError() << ap.provider << " should have a target";
+    if (ap.func.empty() && ap.address == 0)
+      ap.addError() << ap.provider
+                    << " should be attached to a function and/or address";
+    if (!ap.lang.empty() && !is_supported_lang(ap.lang))
+      ap.addError() << "unsupported language type: " << ap.lang;
+
+    if (ap.provider == "uretprobe" && ap.func_offset != 0)
+      ap.addError() << "uretprobes can not be attached to a function offset";
+
+    auto get_paths = [&]() -> Result<std::vector<std::string>> {
+      const auto pid = bpftrace_.pid();
+      if (ap.target == "*") {
+        if (pid.has_value())
+          return util::get_mapped_paths_for_pid(*pid);
+        else
+          return util::get_mapped_paths_for_running_pids();
+      } else {
+        return util::resolve_binary_path(ap.target, pid);
+      }
+    };
+    auto paths = get_paths();
+    if (!paths) {
+      // There was an error during path resolution.
+      ap.addError() << "error finding uprobe target: " << paths.takeError();
+    } else {
+      switch (paths->size()) {
+        case 0:
+          ap.addError() << "uprobe target file '" << ap.target
+                        << "' does not exist or is not executable";
+          break;
+        case 1:
+          // Replace the glob at this stage only if this is *not* a wildcard,
+          // otherwise we rely on the probe matcher. This is not going through
+          // any interfaces that can be properly mocked.
+          if (ap.target.find("*") == std::string::npos)
+            ap.target = paths->front();
+          break;
+        default:
+          // If we are doing a PATH lookup (ie not glob), we follow shell
+          // behavior and take the first match.
+          // Otherwise we keep the target with glob, it will be expanded later
+          if (ap.target.find("*") == std::string::npos) {
+            ap.addWarning() << "attaching to uprobe target file '"
+                            << paths->front() << "' but matched "
+                            << std::to_string(paths->size()) << " binaries";
+            ap.target = paths->front();
+          }
+      }
+    }
+  } else if (ap.provider == "usdt") {
+    bpftrace_.has_usdt_ = true;
+    if (ap.func.empty())
+      ap.addError() << "usdt probe must have a target function or wildcard";
+
+    if (!ap.target.empty() &&
+        !(bpftrace_.pid().has_value() && util::has_wildcard(ap.target))) {
+      auto paths = util::resolve_binary_path(ap.target, bpftrace_.pid());
+      switch (paths.size()) {
+        case 0:
+          ap.addError() << "usdt target file '" << ap.target
+                        << "' does not exist or is not executable";
+          break;
+        case 1:
+          // See uprobe, above.
+          if (ap.target.find("*") == std::string::npos)
+            ap.target = paths.front();
+          break;
+        default:
+          // See uprobe, above.
+          if (ap.target.find("*") == std::string::npos) {
+            ap.addWarning() << "attaching to usdt target file '"
+                            << paths.front() << "' but matched "
+                            << std::to_string(paths.size()) << " binaries";
+            ap.target = paths.front();
+          }
+      }
+    }
+
+    const auto pid = bpftrace_.pid();
+    if (pid.has_value()) {
+      USDTHelper::probes_for_pid(*pid, bpftrace_.feature_->has_uprobe_multi());
+    } else if (ap.target == "*") {
+      USDTHelper::probes_for_all_pids(bpftrace_.feature_->has_uprobe_multi());
+    } else if (!ap.target.empty()) {
+      for (auto &path : util::resolve_binary_path(ap.target))
+        USDTHelper::probes_for_path(path,
+                                    bpftrace_.feature_->has_uprobe_multi());
+    } else {
+      ap.addError() << "usdt probe must specify at least path or pid to "
+                       "probe. To target "
+                       "all paths/pids set the path to '*'.";
+    }
+  } else if (ap.provider == "tracepoint") {
+    if (ap.target.empty() || ap.func.empty())
+      ap.addError() << "tracepoint probe must have a target";
+  } else if (ap.provider == "rawtracepoint") {
+    if (ap.func.empty())
+      ap.addError() << "rawtracepoint should be attached to a function";
+
+    if (!listing_ && !bpftrace_.has_btf_data()) {
+      ap.addError() << "rawtracepoints require kernel BTF. Try using a "
+                       "'tracepoint' instead.";
+    }
+
+  } else if (ap.provider == "profile") {
+    if (ap.target.empty())
+      ap.addError() << "profile probe must have unit of time";
+    else if (!listing_) {
+      if (!TIME_UNITS.contains(ap.target))
+        ap.addError() << ap.target << " is not an accepted unit of time";
+      if (!ap.func.empty())
+        ap.addError() << "profile probe must have an integer frequency";
+      else if (ap.freq <= 0)
+        ap.addError() << "profile frequency should be a positive integer";
+    }
+  } else if (ap.provider == "interval") {
+    if (ap.target.empty())
+      ap.addError() << "interval probe must have unit of time";
+    else if (!listing_) {
+      if (!TIME_UNITS.contains(ap.target))
+        ap.addError() << ap.target << " is not an accepted unit of time";
+      if (!ap.func.empty())
+        ap.addError() << "interval probe must have an integer frequency";
+      else if (ap.freq <= 0)
+        ap.addError() << "interval frequency should be a positive integer";
+    }
+  } else if (ap.provider == "software") {
+    if (ap.target.empty())
+      ap.addError() << "software probe must have a software event name";
+    else {
+      if (!util::has_wildcard(ap.target) && !ap.ignore_invalid) {
+        bool found = false;
+        for (const auto &probeListItem : SW_PROBE_LIST) {
+          if (ap.target == probeListItem.path ||
+              (!probeListItem.alias.empty() &&
+               ap.target == probeListItem.alias)) {
+            found = true;
+            break;
+          }
+        }
+        if (!found)
+          ap.addError() << ap.target << " is not a software probe";
+      } else if (!listing_) {
+        ap.addError() << "wildcards are not allowed for software probe type";
+      }
+    }
+    if (!ap.func.empty())
+      ap.addError() << "software probe can only have an integer count";
+    else if (ap.freq < 0)
+      ap.addError() << "software count should be a positive integer";
+  } else if (ap.provider == "watchpoint" || ap.provider == "asyncwatchpoint") {
+    if (!ap.func.empty()) {
+      if (!bpftrace_.pid().has_value() && !has_child_)
+        ap.addError() << "-p PID or -c CMD required for watchpoint";
+
+      if (ap.address >= static_cast<uint64_t>(arch::Host::arguments().size()))
+        ap.addError() << arch::Host::Machine << " doesn't support arg"
+                      << ap.address;
+    } else if (ap.provider == "asyncwatchpoint")
+      ap.addError() << ap.provider << " requires a function name";
+    else if (!ap.address)
+      ap.addError() << "watchpoint must be attached to a non-zero address";
+    if (ap.len != 1 && ap.len != 2 && ap.len != 4 && ap.len != 8)
+      ap.addError() << "watchpoint length must be one of (1,2,4,8)";
+    if (ap.mode.empty())
+      ap.addError() << "watchpoint mode must be combination of (r,w,x)";
+    std::ranges::sort(ap.mode);
+    for (const char c : ap.mode) {
+      if (c != 'r' && c != 'w' && c != 'x')
+        ap.addError() << "watchpoint mode must be combination of (r,w,x)";
+    }
+    for (size_t i = 1; i < ap.mode.size(); ++i) {
+      if (ap.mode[i - 1] == ap.mode[i])
+        ap.addError() << "watchpoint modes may not be duplicated";
+    }
+    const auto &modes = arch::Host::watchpoint_modes();
+    if (!modes.contains(ap.mode)) {
+      if (modes.empty()) {
+        // There are no valid modes.
+        ap.addError() << "watchpoints not supported";
+      } else {
+        // Build a suitable error with hint.
+        auto &err = ap.addError();
+        err << "invalid watchpoint mode: " << ap.mode;
+        err.addHint() << "supported modes: "
+                      << util::str_join(std::vector(modes.begin(), modes.end()),
+                                        ",");
+      }
+    }
+  } else if (ap.provider == "hardware") {
+    if (ap.target.empty())
+      ap.addError() << "hardware probe must have a hardware event name";
+    else {
+      if (!util::has_wildcard(ap.target) && !ap.ignore_invalid) {
+        bool found = false;
+        for (const auto &probeListItem : HW_PROBE_LIST) {
+          if (ap.target == probeListItem.path ||
+              (!probeListItem.alias.empty() &&
+               ap.target == probeListItem.alias)) {
+            found = true;
+            break;
+          }
+        }
+        if (!found)
+          ap.addError() << ap.target + " is not a hardware probe";
+      } else if (!listing_) {
+        ap.addError() << "wildcards are not allowed for hardware probe type";
+      }
+    }
+    if (!ap.func.empty())
+      ap.addError() << "hardware probe can only have an integer count";
+    else if (ap.freq < 0)
+      ap.addError() << "hardware frequency should be a positive integer";
+  } else if (ap.provider == "begin" || ap.provider == "end") {
+    if (!ap.target.empty() || !ap.func.empty())
+      ap.addError() << "begin/end probes should not have a target";
+    if (ap.provider == "begin") {
+      if (has_begin_probe_)
+        ap.addError() << "More than one begin probe defined";
+      has_begin_probe_ = true;
+    }
+    if (ap.provider == "end") {
+      if (has_end_probe_)
+        ap.addError() << "More than one end probe defined";
+      has_end_probe_ = true;
+    }
+  } else if (ap.provider == "self") {
+    if (ap.target == "signal") {
+      if (!SIGNALS.contains(ap.func))
+        ap.addError() << ap.func << " is not a supported signal";
+      return;
+    }
+    ap.addError() << ap.target << " is not a supported trigger";
+  } else if (ap.provider == "bench") {
+    if (ap.target.empty())
+      ap.addError() << "bench probes must have a name";
+    auto it = benchmark_locs_.find(ap.target);
+
+    if (it != benchmark_locs_.end()) {
+      auto &err = ap.addError();
+      err << "\"" + ap.target + "\""
+          << " was used as the name for more than one BENCH probe";
+      err.addContext(it->second) << "this is the other instance";
+    }
+
+    benchmark_locs_.emplace(ap.target, ap.loc);
+  } else if (ap.provider == "fentry" || ap.provider == "fexit") {
+    if (ap.func.empty())
+      ap.addError() << "fentry/fexit should specify a function";
+  } else if (ap.provider == "iter") {
+    if (!listing_ && bpftrace_.btf_->has_data() &&
+        !bpftrace_.btf_->get_all_iters().contains(ap.func)) {
+      ap.addError() << "iter " << ap.func
+                    << " not available for your kernel version.";
+    }
+
+    if (ap.func.empty())
+      ap.addError() << "iter should specify a iterator's name";
+  } else {
+    ap.addError() << "Invalid provider: '" << ap.provider << "'";
+  }
+}
 
 AttachPointParser::State AttachPointParser::argument_count_error(
     int expected,
@@ -797,10 +1102,24 @@ AttachPointParser::State AttachPointParser::raw_tracepoint_parser()
 // Note: listing changes the parsing semantics for attach points
 Pass CreateParseAttachpointsPass(bool listing)
 {
-  return Pass::create("attachpoints", [listing](ASTContext &ast, BPFtrace &b) {
-    AttachPointParser ap_parser(ast, b, listing);
-    ap_parser.parse();
-  });
+  return Pass::create("parse-attachpoints",
+                      [listing](ASTContext &ast, BPFtrace &b) {
+                        AttachPointParser ap_parser(ast, b, listing);
+                        ap_parser.parse();
+                      });
+}
+
+Pass CreateCheckAttachpointsPass(bool listing)
+{
+  return Pass::create("check-attachpoints",
+                      [listing](ASTContext &ast, BPFtrace &b) {
+                        AttachPointChecker ap_checker(ast,
+                                                      b,
+                                                      listing,
+                                                      !b.cmd_.empty() ||
+                                                          b.child_ != nullptr);
+                        ap_checker.visit(*ast.root);
+                      });
 }
 
 } // namespace bpftrace::ast
