+++ src/ast/passes/semantic_analyser.cpp
@@ -147,15 +147,13 @@ public:
                             MapMetadata &map_metadata,
                             NamedParamDefaults &named_param_defaults,
                             TypeMetadata &type_metadata,
-                            bool has_child = true,
-                            bool listing = false)
+                            bool has_child = true)
       : ctx_(ctx),
         bpftrace_(bpftrace),
         c_definitions_(c_definitions),
         map_metadata_(map_metadata),
         named_param_defaults_(named_param_defaults),
         type_metadata_(type_metadata),
-        listing_(listing),
         has_child_(has_child)
   {
   }
@@ -194,7 +192,6 @@ public:
   void visit(AssignVarStatement &assignment);
   void visit(VarDeclStatement &decl);
   void visit(Unroll &unroll);
-  void visit(AttachPoint &ap);
   void visit(Probe &probe);
   void visit(BlockExpr &block);
   void visit(Subprog &subprog);
@@ -208,7 +205,6 @@ private:
   MapMetadata &map_metadata_;
   NamedParamDefaults &named_param_defaults_;
   TypeMetadata &type_metadata_;
-  bool listing_;
 
   bool is_final_pass() const;
   bool is_first_pass() const;
@@ -300,10 +296,7 @@ private:
   std::map<std::string, bpf_map_type> bpf_map_type_;
 
   uint32_t loop_depth_ = 0;
-  bool has_begin_probe_ = false;
-  bool has_end_probe_ = false;
   bool has_child_ = false;
-  std::unordered_map<std::string, Location> benchmark_locs_;
 };
 
 } // namespace
@@ -3840,291 +3833,6 @@ void SemanticAnalyser::visit(VarDeclStatement &decl)
   variable_decls_[scope_stack_.back()].insert({ var_ident, decl });
 }
 
-void SemanticAnalyser::visit(AttachPoint &ap)
-{
-  if (ap.provider == "kprobe" || ap.provider == "kretprobe") {
-    if (ap.func.empty())
-      ap.addError() << "kprobes should be attached to a function";
-    if (is_final_pass()) {
-      // Warn if user tries to attach to a non-traceable function
-      if (bpftrace_.config_->missing_probes != ConfigMissingProbes::ignore &&
-          !util::has_wildcard(ap.func) &&
-          !bpftrace_.is_traceable_func(ap.func)) {
-        ap.addWarning() << ap.func
-                        << " is not traceable (either non-existing, inlined, "
-                           "or marked as "
-                           "\"notrace\"); attaching to it will likely fail";
-      }
-    }
-  } else if (ap.provider == "uprobe" || ap.provider == "uretprobe") {
-    if (ap.target.empty())
-      ap.addError() << ap.provider << " should have a target";
-    if (ap.func.empty() && ap.address == 0)
-      ap.addError() << ap.provider
-                    << " should be attached to a function and/or address";
-    if (!ap.lang.empty() && !is_supported_lang(ap.lang))
-      ap.addError() << "unsupported language type: " << ap.lang;
-
-    if (ap.provider == "uretprobe" && ap.func_offset != 0)
-      ap.addError() << "uretprobes can not be attached to a function offset";
-
-    auto get_paths = [&]() -> Result<std::vector<std::string>> {
-      const auto pid = bpftrace_.pid();
-      if (ap.target == "*") {
-        if (pid.has_value())
-          return util::get_mapped_paths_for_pid(*pid);
-        else
-          return util::get_mapped_paths_for_running_pids();
-      } else {
-        return util::resolve_binary_path(ap.target, pid);
-      }
-    };
-    auto paths = get_paths();
-    if (!paths) {
-      // There was an error during path resolution.
-      ap.addError() << "error finding uprobe target: " << paths.takeError();
-    } else {
-      switch (paths->size()) {
-        case 0:
-          ap.addError() << "uprobe target file '" << ap.target
-                        << "' does not exist or is not executable";
-          break;
-        case 1:
-          // Replace the glob at this stage only if this is *not* a wildcard,
-          // otherwise we rely on the probe matcher. This is not going through
-          // any interfaces that can be properly mocked.
-          if (ap.target.find("*") == std::string::npos)
-            ap.target = paths->front();
-          break;
-        default:
-          // If we are doing a PATH lookup (ie not glob), we follow shell
-          // behavior and take the first match.
-          // Otherwise we keep the target with glob, it will be expanded later
-          if (ap.target.find("*") == std::string::npos) {
-            ap.addWarning() << "attaching to uprobe target file '"
-                            << paths->front() << "' but matched "
-                            << std::to_string(paths->size()) << " binaries";
-            ap.target = paths->front();
-          }
-      }
-    }
-  } else if (ap.provider == "usdt") {
-    bpftrace_.has_usdt_ = true;
-    if (ap.func.empty())
-      ap.addError() << "usdt probe must have a target function or wildcard";
-
-    if (!ap.target.empty() &&
-        !(bpftrace_.pid().has_value() && util::has_wildcard(ap.target))) {
-      auto paths = util::resolve_binary_path(ap.target, bpftrace_.pid());
-      switch (paths.size()) {
-        case 0:
-          ap.addError() << "usdt target file '" << ap.target
-                        << "' does not exist or is not executable";
-          break;
-        case 1:
-          // See uprobe, above.
-          if (ap.target.find("*") == std::string::npos)
-            ap.target = paths.front();
-          break;
-        default:
-          // See uprobe, above.
-          if (ap.target.find("*") == std::string::npos) {
-            ap.addWarning() << "attaching to usdt target file '"
-                            << paths.front() << "' but matched "
-                            << std::to_string(paths.size()) << " binaries";
-            ap.target = paths.front();
-          }
-      }
-    }
-
-    const auto pid = bpftrace_.pid();
-    if (pid.has_value()) {
-      USDTHelper::probes_for_pid(*pid, bpftrace_.feature_->has_uprobe_multi());
-    } else if (ap.target == "*") {
-      USDTHelper::probes_for_all_pids(bpftrace_.feature_->has_uprobe_multi());
-    } else if (!ap.target.empty()) {
-      for (auto &path : util::resolve_binary_path(ap.target))
-        USDTHelper::probes_for_path(path,
-                                    bpftrace_.feature_->has_uprobe_multi());
-    } else {
-      ap.addError() << "usdt probe must specify at least path or pid to "
-                       "probe. To target "
-                       "all paths/pids set the path to '*'.";
-    }
-  } else if (ap.provider == "tracepoint") {
-    if (ap.target.empty() || ap.func.empty())
-      ap.addError() << "tracepoint probe must have a target";
-  } else if (ap.provider == "rawtracepoint") {
-    if (ap.func.empty())
-      ap.addError() << "rawtracepoint should be attached to a function";
-
-    if (!listing_ && !bpftrace_.has_btf_data()) {
-      ap.addError() << "rawtracepoints require kernel BTF. Try using a "
-                       "'tracepoint' instead.";
-    }
-
-  } else if (ap.provider == "profile") {
-    if (ap.target.empty())
-      ap.addError() << "profile probe must have unit of time";
-    else if (!listing_) {
-      if (!TIME_UNITS.contains(ap.target))
-        ap.addError() << ap.target << " is not an accepted unit of time";
-      if (!ap.func.empty())
-        ap.addError() << "profile probe must have an integer frequency";
-      else if (ap.freq <= 0)
-        ap.addError() << "profile frequency should be a positive integer";
-    }
-  } else if (ap.provider == "interval") {
-    if (ap.target.empty())
-      ap.addError() << "interval probe must have unit of time";
-    else if (!listing_) {
-      if (!TIME_UNITS.contains(ap.target))
-        ap.addError() << ap.target << " is not an accepted unit of time";
-      if (!ap.func.empty())
-        ap.addError() << "interval probe must have an integer frequency";
-      else if (ap.freq <= 0)
-        ap.addError() << "interval frequency should be a positive integer";
-    }
-  } else if (ap.provider == "software") {
-    if (ap.target.empty())
-      ap.addError() << "software probe must have a software event name";
-    else {
-      if (!util::has_wildcard(ap.target) && !ap.ignore_invalid) {
-        bool found = false;
-        for (const auto &probeListItem : SW_PROBE_LIST) {
-          if (ap.target == probeListItem.path ||
-              (!probeListItem.alias.empty() &&
-               ap.target == probeListItem.alias)) {
-            found = true;
-            break;
-          }
-        }
-        if (!found)
-          ap.addError() << ap.target << " is not a software probe";
-      } else if (!listing_) {
-        ap.addError() << "wildcards are not allowed for hardware probe type";
-      }
-    }
-    if (!ap.func.empty())
-      ap.addError() << "software probe can only have an integer count";
-    else if (ap.freq < 0)
-      ap.addError() << "software count should be a positive integer";
-  } else if (ap.provider == "watchpoint" || ap.provider == "asyncwatchpoint") {
-    if (!ap.func.empty()) {
-      if (!bpftrace_.pid().has_value() && !has_child_)
-        ap.addError() << "-p PID or -c CMD required for watchpoint";
-
-      if (ap.address >= static_cast<uint64_t>(arch::Host::arguments().size()))
-        ap.addError() << arch::Host::Machine << " doesn't support arg"
-                      << ap.address;
-    } else if (ap.provider == "asyncwatchpoint")
-      ap.addError() << ap.provider << " requires a function name";
-    else if (!ap.address)
-      ap.addError() << "watchpoint must be attached to a non-zero address";
-    if (ap.len != 1 && ap.len != 2 && ap.len != 4 && ap.len != 8)
-      ap.addError() << "watchpoint length must be one of (1,2,4,8)";
-    if (ap.mode.empty())
-      ap.addError() << "watchpoint mode must be combination of (r,w,x)";
-    std::ranges::sort(ap.mode);
-    for (const char c : ap.mode) {
-      if (c != 'r' && c != 'w' && c != 'x')
-        ap.addError() << "watchpoint mode must be combination of (r,w,x)";
-    }
-    for (size_t i = 1; i < ap.mode.size(); ++i) {
-      if (ap.mode[i - 1] == ap.mode[i])
-        ap.addError() << "watchpoint modes may not be duplicated";
-    }
-    const auto &modes = arch::Host::watchpoint_modes();
-    if (!modes.contains(ap.mode)) {
-      if (modes.empty()) {
-        // There are no valid modes.
-        ap.addError() << "watchpoints not supported";
-      } else {
-        // Build a suitable error with hint.
-        auto &err = ap.addError();
-        err << "invalid watchpoint mode: " << ap.mode;
-        err.addHint() << "supported modes: "
-                      << util::str_join(std::vector(modes.begin(), modes.end()),
-                                        ",");
-      }
-    }
-  } else if (ap.provider == "hardware") {
-    if (ap.target.empty())
-      ap.addError() << "hardware probe must have a hardware event name";
-    else {
-      if (!util::has_wildcard(ap.target) && !ap.ignore_invalid) {
-        bool found = false;
-        for (const auto &probeListItem : HW_PROBE_LIST) {
-          if (ap.target == probeListItem.path ||
-              (!probeListItem.alias.empty() &&
-               ap.target == probeListItem.alias)) {
-            found = true;
-            break;
-          }
-        }
-        if (!found)
-          ap.addError() << ap.target + " is not a hardware probe";
-      } else if (!listing_) {
-        ap.addError() << "wildcards are not allowed for hardware probe type";
-      }
-    }
-    if (!ap.func.empty())
-      ap.addError() << "hardware probe can only have an integer count";
-    else if (ap.freq < 0)
-      ap.addError() << "hardware frequency should be a positive integer";
-  } else if (ap.provider == "begin" || ap.provider == "end") {
-    if (!ap.target.empty() || !ap.func.empty())
-      ap.addError() << "begin/end probes should not have a target";
-    if (is_final_pass()) {
-      if (ap.provider == "begin") {
-        if (has_begin_probe_)
-          ap.addError() << "More than one begin probe defined";
-        has_begin_probe_ = true;
-      }
-      if (ap.provider == "end") {
-        if (has_end_probe_)
-          ap.addError() << "More than one end probe defined";
-        has_end_probe_ = true;
-      }
-    }
-  } else if (ap.provider == "self") {
-    if (ap.target == "signal") {
-      if (!SIGNALS.contains(ap.func))
-        ap.addError() << ap.func << " is not a supported signal";
-      return;
-    }
-    ap.addError() << ap.target << " is not a supported trigger";
-  } else if (ap.provider == "bench") {
-    if (ap.target.empty())
-      ap.addError() << "bench probes must have a name";
-    if (is_final_pass()) {
-      auto it = benchmark_locs_.find(ap.target);
-
-      if (it != benchmark_locs_.end()) {
-        auto &err = ap.addError();
-        err << "\"" + ap.target + "\""
-            << " was used as the name for more than one BENCH probe";
-        err.addContext(it->second) << "this is the other instance";
-      }
-
-      benchmark_locs_.emplace(ap.target, ap.loc);
-    }
-  } else if (ap.provider == "fentry" || ap.provider == "fexit") {
-    if (ap.func.empty())
-      ap.addError() << "fentry/fexit should specify a function";
-  } else if (ap.provider == "iter") {
-    if (!listing_ && !bpftrace_.btf_->get_all_iters().contains(ap.func)) {
-      ap.addError() << "iter " << ap.func
-                    << " not available for your kernel version.";
-    }
-
-    if (ap.func.empty())
-      ap.addError() << "iter should specify a iterator's name";
-  } else {
-    ap.addError() << "Invalid provider: '" << ap.provider << "'";
-  }
-}
-
 void SemanticAnalyser::visit(BlockExpr &block)
 {
   scope_stack_.push_back(&block);
@@ -4725,22 +4433,21 @@ void SemanticAnalyser::resolve_struct_type(SizedType &type, Node &node)
   }
 }
 
-Pass CreateSemanticPass(bool listing)
+Pass CreateSemanticPass()
 {
-  auto fn = [listing](ASTContext &ast,
-                      BPFtrace &b,
-                      CDefinitions &c_definitions,
-                      MapMetadata &mm,
-                      NamedParamDefaults &named_param_defaults,
-                      TypeMetadata &types) {
+  auto fn = [](ASTContext &ast,
+               BPFtrace &b,
+               CDefinitions &c_definitions,
+               MapMetadata &mm,
+               NamedParamDefaults &named_param_defaults,
+               TypeMetadata &types) {
     SemanticAnalyser semantics(ast,
                                b,
                                c_definitions,
                                mm,
                                named_param_defaults,
                                types,
-                               !b.cmd_.empty() || b.child_ != nullptr,
-                               listing);
+                               !b.cmd_.empty() || b.child_ != nullptr);
     semantics.analyse();
   };
