+++ tests/CMakeLists.txt
@@ -28,7 +28,7 @@ endif()
 add_executable(bpftrace_test
   arch.cpp
   async_action.cpp
-  attachpoint_parser.cpp
+  attachpoint_passes.cpp
   bpfbytecode.cpp
   bpftrace.cpp
   btf.cpp
