+++ tests/bpftrace.cpp
@@ -2,7 +2,7 @@
 #include <cstdint>
 #include <cstring>
 
-#include "ast/attachpoint_parser.h"
+#include "ast/passes/attachpoint_passes.h"
 #include "ast/passes/clang_parser.h"
 #include "ast/passes/codegen_llvm.h"
 #include "ast/passes/field_analyser.h"
@@ -66,6 +66,7 @@ static auto parse_probe(const std::string &str,
                 .put(no_types)
                 .add(CreateParsePass())
                 .add(ast::CreateParseAttachpointsPass())
+                .add(ast::CreateCheckAttachpointsPass())
                 .add(ast::CreateProbeExpansionPass())
                 .add(ast::CreateMacroExpansionPass())
                 .add(ast::CreateFieldAnalyserPass())
