+++ tests/clang_parser.cpp
@@ -1,6 +1,6 @@
 #include <llvm/Config/llvm-config.h>
 
-#include "ast/attachpoint_parser.h"
+#include "ast/passes/attachpoint_passes.h"
 #include "ast/passes/clang_parser.h"
 #include "ast/passes/field_analyser.h"
 #include "ast/passes/probe_expansion.h"
