+++ tests/codegen/regression.cpp
@@ -23,6 +23,7 @@ TEST(codegen, regression_957)
                 .add(ast::CreateResolveImportsPass())
                 .add(ast::CreateImportInternalScriptsPass())
                 .add(ast::CreateParseAttachpointsPass())
+                .add(ast::CreateCheckAttachpointsPass())
                 .add(ast::CreateMacroExpansionPass())
                 .add(ast::CreateProbeExpansionPass())
                 .add(ast::CreateFieldAnalyserPass())
