+++ tests/pid_filter_pass.cpp
@@ -1,5 +1,5 @@
 #include "ast/passes/pid_filter_pass.h"
-#include "ast/attachpoint_parser.h"
+#include "ast/passes/attachpoint_passes.h"
 #include "ast/passes/field_analyser.h"
 #include "ast/passes/printer.h"
 #include "ast/passes/probe_expansion.h"
