+++ tests/portability_analyser.cpp
@@ -1,5 +1,5 @@
 #include "ast/passes/portability_analyser.h"
-#include "ast/attachpoint_parser.h"
+#include "ast/passes/attachpoint_passes.h"
 #include "ast/passes/field_analyser.h"
 #include "ast/passes/macro_expansion.h"
 #include "ast/passes/map_sugar.h"
@@ -33,6 +33,7 @@ void test(BPFtrace &bpftrace, const std::string &input, int expected_result = 0)
                 .put(no_types)
                 .add(CreateParsePass())
                 .add(ast::CreateParseAttachpointsPass())
+                .add(ast::CreateCheckAttachpointsPass())
                 .add(ast::CreateProbeExpansionPass())
                 .add(ast::CreateMacroExpansionPass())
                 .add(ast::CreateMapSugarPass())
